more cleanup of nes_cpu: NES_BASECYCLES is not constant, changing capitalization Use Exec instead of exec for init missed a few things in that INT64 conversion from last push

+++ xgm/xtypes.h
@@ -18,7 +18,7 @@ typedef unsigned char UINT8;
 /** 16bit•„†‚È‚µ‚Ì®”Œ^ */
 typedef unsigned short UINT16;
 /** 32bit•„†‚È‚µ‚Ì®”Œ^ */
-typedef unsigned long UINT32;
+typedef unsigned int UINT32;
 /** 64bit unsigned */
 typedef unsigned long long UINT64;
 /** 8bit•„†•t‚«®”Œ^ */
@@ -26,7 +26,7 @@ typedef signed char INT8;
 /** 16bit•„†•t‚«®”Œ^ */
 typedef signed short INT16;
 /** 32bit•„†•t‚«®”Œ^ */
-typedef signed long INT32;
+typedef signed int INT32;
 /** 64bit signed */
 typedef signed long long INT64;
 }
